+++ zakura-consensus/src/lib.rs
@@ -49,7 +49,10 @@ pub use block::check::difficulty_is_valid;
 #[cfg(any(test, feature = "proptest-impl"))]
 pub use checkpoint::CheckpointVerifier;
 
-pub use block::{subsidy::funding_stream_address, Request, VerifyBlockError, MAX_BLOCK_SIGOPS};
+pub use block::{
+    subsidy::funding_stream_address, Request, SemanticBlockVerifier, VerifyBlockError,
+    MAX_BLOCK_SIGOPS,
+};
 pub use checkpoint::{VerifyCheckpointError, MAX_CHECKPOINT_BYTE_COUNT, MAX_CHECKPOINT_HEIGHT_GAP};
 pub use config::Config;
 pub use error::BlockError;